Separate WhenAll implementation from interface
